Apply patch package to fix build on Xcode 14.3 #10
I found a bug when building the project with Xcode 14.3 and above: "Use of bitwise '|' with boolean operands". You can see more details on this issue. Unfortunately, the fix was applied only on version 0.69 and above. As our project uses 0.68.1 we have two options:

Update to a new RN version
Apply a patch package as per this stack overflow issue.
I'm implementing the second one because we don't have time to do a major update, but as future work would be great to see the project on the latest RN version. Migrate from 0.68.1 to 0.72.4